+++ packages/tools/src/stateManagement/segmentation/getCurrentLabelmapImageIdForViewport.ts
@@ -1,0 +1,19 @@
+import { defaultSegmentationStateManager } from './SegmentationStateManager';
+
+/**
+ * Retrieves the labelmap image IDs for a specific viewport and segmentation representation.
+ *
+ * @param viewportId - The ID of the viewport.
+ * @param segmentationId -  The ID of the segmentation.
+ * @returns An array of labelmap image IDs.
+ */
+export function getCurrentLabelmapImageIdForViewport(
+  viewportId: string,
+  segmentationId?: string
+) {
+  const segmentationStateManager = defaultSegmentationStateManager;
+  return segmentationStateManager.getCurrentLabelmapImageIdForViewport(
+    viewportId,
+    segmentationId
+  );
+}
