+++ packages/tools/src/stateManagement/segmentation/getNextColorLUTIndex.ts
@@ -1,0 +1,6 @@
+import { defaultSegmentationStateManager } from './SegmentationStateManager';
+
+export function getNextColorLUTIndex(): number {
+  const segmentationStateManager = defaultSegmentationStateManager;
+  return segmentationStateManager.getNextColorLUTIndex();
+}
