+++ packages/tools/src/stateManagement/segmentation/getSegmentationRepresentation.ts
@@ -1,0 +1,17 @@
+import type { SegmentationRepresentation } from '../../types/SegmentationStateTypes';
+import { defaultSegmentationStateManager } from './SegmentationStateManager';
+
+/**
+ * Finds a segmentation representation by its UID.
+ *
+ * @param segmentationRepresentationUID - The UID of the segmentation representation to find.
+ * @returns The found segmentation representation, or undefined if not found.
+ */
+export function getSegmentationRepresentation(
+  segmentationRepresentationUID: string
+): SegmentationRepresentation | undefined {
+  const segmentationStateManager = defaultSegmentationStateManager;
+  return segmentationStateManager.getSegmentationRepresentation(
+    segmentationRepresentationUID
+  );
+}
