+++ packages/tools/src/stateManagement/segmentation/updateLabelmapSegmentationImageReferences.ts
@@ -1,0 +1,12 @@
+import { defaultSegmentationStateManager } from './SegmentationStateManager';
+
+export function updateLabelmapSegmentationImageReferences(
+  viewportId: string,
+  segmentationId: string
+) {
+  const segmentationStateManager = defaultSegmentationStateManager;
+  segmentationStateManager.updateLabelmapSegmentationImageReferences(
+    viewportId,
+    segmentationId
+  );
+}
