--- src/stateManagement/segmentation/functions/addSegmentation.ts
@@ -1,1 +1,0 @@
-
