--- src/stateManagement/segmentation/functions/getDefaultSegmentationStateManager.ts
@@ -1,1 +1,0 @@
-
