--- src/stateManagement/segmentation/functions/getSegmentations.ts
@@ -1,1 +1,0 @@
-
