--- src/stateManagement/segmentation/index.ts
@@ -1,1 +1,0 @@
-
